+++ cms/resources/views/frontend/blog/index.blade.php
@@ -1,4 +1,4 @@
-@extends('frontend.layout')
+@extends($themeRuntime->view('layout'))
 
 @section('content')
     <section class="pulse-blog-hero">