+++ cms/resources/views/frontend/page.blade.php
@@ -1,4 +1,4 @@
-@extends('frontend.layout')
+@extends($themeRuntime->view('layout'))
 
 @section('content')
     @if ($builder)